data type optymalization

+++ core/src/main/java/com/astro/core/engine/base/CameraManager.java
@@ -28,7 +28,7 @@ public class CameraManager {
      */
     @Setter
     @Value("${renderer.pixel.per.meter}")
-    protected int pixelPerMeter;
+    protected short pixelPerMeter;
 
     @Getter
     @Autowired
